I added a function for validating clusters
I added a function to validate clusters. Three methods are compared and
results for different clustering methods and for different numbers of
clusters are written into a csv file.
A small hitch is that agreement between clustering methods is
established by comparing the cluster number that each icon gets. I have
to check whether it could be that two clusters are identical but simply
have different numbers.